+++ binance/__init__.py
@@ -3,3 +3,9 @@
 .. moduleauthor:: Sam McHardy
 
 """
+
+__version__ = '1.0.1'
+
+from binance.client import Client, AsyncClient  # noqa
+from binance.depthcache import DepthCacheManager, OptionsDepthCacheManager  # noqa
+from binance.streams import BinanceSocketManager  # noqa