+++ tests/test_api_request.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 
 from binance.client import Client
-from binance.exceptions import BinanceAPIException, BinanceRequestException, BinanceWithdrawException
+from binance.exceptions import BinanceAPIException, BinanceRequestException
 import pytest
 import requests_mock
 
@@ -37,14 +37,3 @@ def test_api_exception_invalid_json():
             not_json_str = "<html><body>Error</body></html>"
             m.get('https://api.binance.com/api/v3/time', text=not_json_str, status_code=400)
             client.get_server_time()
-
-
-def test_withdraw_api_exception():
-    """Test Withdraw API response Exception"""
-
-    with pytest.raises(BinanceWithdrawException):
-
-        with requests_mock.mock() as m:
-            json_obj = {"success": False, "msg": "Insufficient funds"}
-            m.register_uri('POST', requests_mock.ANY, json=json_obj, status_code=200)
-            client.withdraw(asset='BTC', address='BTCADDRESS', amount=100)